LayoutTests/fast/canvas/webgl/array-override-set-expected.txt

Test overriding the set method of WebGL array

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=87862 : [v8] Crash after redefining setter on typed array to a number
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS successfullyParsed is true

TEST COMPLETE
